requirements.txt
alembic==1.16.1
annotated-types==0.7.0
bcrypt==4.3.0
blinker==1.9.0
click==8.2.1
dnspython==2.7.0
ecdsa==0.19.1
email-validator==2.2.0
flask==3.1.1
flask-sqlalchemy==3.1.1
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
mako==1.3.10
markupsafe==3.0.2
psycopg2-binary==2.9.10
pyasn1==0.4.8
pydantic==2.11.5
pydantic-core==2.33.2
python-dotenv==1.1.0
python-jose==3.4.0
rsa==4.9.1
six==1.17.0
sqlalchemy==2.0.41
typing-extensions==4.13.2
typing-inspection==0.4.1
werkzeug==3.1.3
